+++ lib/src/_components.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:domino/domino.dart';
 import 'package:domino/helpers.dart' show div;
 
@@ -26,7 +28,7 @@ Element breadcrumb({mods, List items}) {
   );
 }
 
-Element card({mods, String title, content}) {
+Element card({mods, String title, content, List footer}) {
   final children = [];
   List headerChildren;
   if (title != null) {
@@ -44,7 +46,12 @@ Element card({mods, String title, content}) {
     children.add(div([clazz('card-content'), content]));
   }
 
-  // TODO: add footer
+  if (footer != null) {
+    children.add(Element('footer', [
+      clazz('card-footer'),
+      footer,
+    ]));
+  }
 
   return div([clazz('card'), mods, children]);
 }
@@ -65,7 +72,73 @@ Element message({mods, String title, Function onDelete, content}) {
   return Element('article', [clazz('message'), mods, children]);
 }
 
-// TODO: add Dropdown
+int _dropdownSymbolCounter = 0;
+
+class DropdownMenu extends Component {
+  final dynamic trigger;
+  final dynamic mods;
+  final dynamic menu;
+
+  final _symbol = Symbol('bulma.dropdown.${_dropdownSymbolCounter++}');
+  StreamSubscription _subscription;
+  bool isActive = false;
+  bool _catchAllProtected = false;
+
+  DropdownMenu({this.trigger, this.mods, this.menu});
+
+  @override
+  build(BuildContext context) {
+    return div([
+      clazz('dropdown'),
+      _symbol,
+      afterInsert((c) {
+        print('after-insert');
+        _subscription ??= c.node.ownerDocument.onClick.listen((e) {
+          if (isActive && !_catchAllProtected) {
+            isActive = false;
+            context.view.invalidate();
+          }
+          _catchAllProtected = false;
+        }) as StreamSubscription;
+      }),
+      afterRemove((c) {
+        print('after-remove');
+        _subscription?.cancel();
+        _subscription = null;
+      }),
+      if (isActive) Modifier.active,
+      mods,
+      div([
+        clazz('dropdown-trigger'),
+        on('click', () {
+          print('trigger click');
+          isActive = !isActive;
+          _catchAllProtected = true;
+        }),
+        trigger,
+      ]),
+      div([
+        clazz('dropdown-menu'),
+        attr('role', 'menu'),
+        div([clazz('dropdown-content'), menu]),
+      ]),
+    ]);
+  }
+}
+
+Element dropdownItem(content, {Function onClick}) {
+  if (onClick != null) {
+    return Element(
+        'a', [clazz('dropdown-item'), on('click', onClick), content]);
+  } else {
+    return div([clazz('dropdown-item'), content]);
+  }
+}
+
+Element dropdownSep() {
+  return Element('hr', clazz('dropdown-divider'));
+}
+
 // TODO: add Menu
 
 Element modal({header, content, footer, close, bool isActive = false}) {
@@ -127,6 +200,7 @@ Element pagination({mods, previous, next, List items}) {
 
 Element tabs({mods, List items}) {
   return Element('div', [
+    clazz('tabs'),
     mods,
     Element('ul', [
       items.map((item) => Element('li', [item])).toList(),